feat: adding option overwrite

+++ src/commands/keygen/create.ts
@@ -1,9 +1,10 @@
-import { writeFileSync } from "fs";
+import { writeFileSync, existsSync } from "fs";
 import { ethers } from "ethers";
 import { ConfigFileManager } from "../../lib/config/ConfigFileManager";
 
 export interface CreateKeypairOptions {
   output: string;
+  overwrite: boolean;
 }
 
 export class KeypairCreator {
@@ -15,13 +16,22 @@ export class KeypairCreator {
 
   createKeypairAction(options: CreateKeypairOptions) {
     try {
+
+      const outputPath = this.filePathManager.getFilePath(options.output);
+
+      if(existsSync(outputPath) && !options.overwrite) {
+        console.warn(
+          `The file at ${outputPath} already exists. Use the '--overwrite' option to replace it.`
+        );
+        return;
+      }
+
       const wallet = ethers.Wallet.createRandom();
       const keypairData = {
         address: wallet.address,
         privateKey: wallet.privateKey,
       };
 
-      const outputPath = this.filePathManager.getFilePath(options.output);
       writeFileSync(outputPath, JSON.stringify(keypairData, null, 2));
 
       this.filePathManager.writeConfig('keyPairPath', outputPath);
